Add a note about the output completion check
This may cause some issues when the process is marked as complete here.
The downstream processes that are not complete and waiting for another
datum packet will never get a completion stamp from this process and
will hang. So far, processes only have all required or all optional
output ports (excepting the heartbeat port which is handled differently)
and this is not an issue. When new processes crop up that do have
required and optional output ports, this must be addressed.