Import plugins with global visibility
Unfortunately, Python has a weird tangle where modules don't link
libpython.so because if Python is built statically, there is no dynamic
library to link to, initializers and such cannot be satisfied. If Python
is built statically and libpython.so exists, then there can be duplicate
symbols.

If the module loader is loaded from C++ code and its call to
Py_Initialize is the one that actually starts the interpreter, then its
loading of libpython.so is behind a RTLD_LOCAL barrier that prevents
Python modules from having access to libpython.so.