Ignore unconnected ports
When input ports are automatically grabbed from, they need to be checked
for an input edge first. If not, an exception will be thrown when the
grab happens. This usually occurs at shutdown time with optional ports
that were left unconnected, but will happen with any automatic grabbing
(error, empty, flush, etc.).