Bind conversion function in configuration step
Go ahead and determine what flavor of conversion function we will need
in read_video_gravl_process::_configure, and stick it in a
boost::function with the pixel type arguments bound. This saves effort
later and allows us to determine earlier if we are unable to produce the
requested output format (and also matches the exception we were
throwing).