Add note to improve error reporting
Add a \todo that we should at some point write an error handler to try
to extract more useful information about why we were not able to obtain
a resource. Also remove check for bad video pointer (after we've already
dereferenced it, to boot) that should never make it past configuration.