Replace find with count methods
Where a find() call was made only to detect whether an element is part
of a container, the count() method (or std::count) can be used instead.
This saves a variable and makes it easier to tell whether "is a member"
or "is not a member" is being tested (since the comparison to .end() is
backwards).