Add logic to check shared output port logic
A shared output port may only have at most one mutable input port
listening to it. Any number of non-mutable input ports may be connected,
but any mutable ports spoil the whole party. The reasoning is that since
the output data is shared among all readers of the port, if the process
with the mutate port runs first, it can result in the other receivers
from seeing its new data instead of the data the upstream actually
created.