Add tests for loading files with slashes in names
Slashes are allowed in process names, but only by way of either manual
specification (in which case, the guarantees the parsers make are
invalid) or via the process cluster namespacing. As a result, slashes
are allowed in port mappings in clusters and connection specifications.
They are not allowed in cluster or process names when declared in a
pipeline file.