Throw when exporting empty-name processes to DOT
It's an implementation detail that the empty string is used as "root"
level when recursing through the tree to export a pipeline, but an empty
string process is also going to cause issues as a node name (plus it's
not going to show up in the end graph). To fix this, empty strings are
not allowed within exported (at least, to the DOT format) pipelines
anymore.