Setup for mapping configs from loaded clusters
Due to the fact that the cluster configuration format currently
implements map_config with the CONF provider, the tunable machinery
within process_cluster cannot be used (due to this information never
reaching the cluster class).

To fix this, any line which uses the CONF provider to request which is
requesting a parameter from the main cluster by name is used to create a
configuration mapping. Unfortunately, with the way things currently
work, some extra steps must be done to ensure that the cluster won't be
setting any values on the processes which are mapped.

There are some issues with the sanity checking yet, but more information
is required to do much more checking.

Some tests are referenced by name, but have not been committed yet.