Rework Windows tests due to library madness
Since Windows doesn't use rpath, we can't just run the tests from
anywhere. Unfortunately, CMake does not (currently) support generator
expressions in test properties, so we are forced to do some hacks for
the tests.

For non-Python, we output the test binaries to the same place as all the
other binaries and use CTest's default working directory of "where the
binary is" to get the correct behavior.

For Python, we have to change the directory to the correct location.
Since tests do have their arguments expanded for generator expressions,
we pass it as an argument and chdir in the Python code for it.