Add a function to parse tests from the sources
Similar to the Python function, this inspects the source code of the
test file to discover which tests are available. A dummy macro is used
for the test properties to ensure that the property takes (at least) two
arguments. The variable argument macro is required in case a property
has a comma in it.