Removed semantically redundant MAPTK_PLUGIN_ENABLE_* vars
Now we just have one flag that enabled searching for additional
dependencies and building the plugins that require those dependencies.

Original intent of the way things were was that multiple plugins may
require the same dependency so just turning on a dependency would not
automatically enable all plugins that used that dependency. However,
this situation is not the current case and causes confused with current
uses.