Removed unneeded function export labeling that broke windows build
Found code that was not intended to have landed with the addition of the
plugin shell template file that defined expected function export
semantics. Plugin standard registration function is compiled into the same
library as the standard wrapper code, so the plugin side function does not
need to apply (and shouldn't) any import/export semantics.