Add program description to tools infrastructure.
There was no way of describing what a tool does or why I would want to
use it (even though the tools had really cool names).

Tool help should be equivalent to a man page. It's not that complete
yet, but at least there is a way to add more description.