Fixed some formatting issues when printing vector2d
Since vector2d is now Eigen::Matrix<T,2,1>, it is a column vector and
prints (via <<) with a newline between the values.  In some cases we
want to print both values on one line, so we need to call .transpose()
to make a row vector instead.

This was causing problem in both config files and the track files.