Use defaulted destructors
Replace empty inline destructors with defaulted destructors, where
available (retaining the empty destructor otherwise). This potentially
permits additional compiler optimizations, as a defaulted destructor is
not quite the same thing as an empty destructor. (Particularly, a
defaulted destructor may be 'trivial', while an empty destructor is not
permitted to be considered 'trivial'.)