Reoranize lens distortion functions use them in a templated cost function
Move the lens distortion functions to a new header.
Make the lens distortion functions static members functions of a class
that also provides a static constant for the number of parameters.
Template the Ceres distortion cost function over distortation class to
remove the many very similar cost functions classes.
Create a factory function to generate the appropriate Ceres cost
function based on enum type, and use this in bundle_adjust.