Modify module loading path handling.
Add optional set of directories from CMake input. This allows a set of
search directories to be specified at CMake time.

Pass configurable data to the APM through a configured include file. This
gets around the problem of using ';' as a path separator in CMake. It
also simplified the code in the APM. Now there is only one configured
list of directories and the set from the environment.