Allow loading camera from image path and directory
Add an overload of read_krtd_file that takes an image file path and
camera directory, like read_krtd_files, but for reading only one camera.
Modify read_krtd_files to use this.

This allows users to load cameras in a manner similar to read_krtd_files
in cases where, for whatever reason, the user wants or needs to do the
iteration over the image list themselves.