Add helper to iterate over sparse matrices
Add a helper class that turns an Eigen::SparseMatrix into something that
we can trivially iterate over (e.g. with VITAL_FOREACH) in order to
visit all non-zero cells.

The choice of "enumerate" for the helper function is from the dictionary
definition, "to specify each member of a sequence individually in
incrementing order" (https://en.wiktionary.org/wiki/enumerate).