Add unit test of sparse matrix enumeration
Add a unit test for the sparse matrix enumeration helper that was added
in the previous commit. This sets up some sparse matrices and performs
takes a sum and sum-product of the matrices [with another "matrix"], in
the manner of image filtering by convolution kernel.

And... it fails, due to a bug in the enumeration helper. This, kids, is
why we write unit tests!