Fix sparse matrix enumeration
Fix bugs in the sparse matrix enumeration helper. For some cases, the
outer vector may contain empty inner vectors, which require us to
immediately check the inner iterator for validity (as it can be
constructed invalid). Modify logic (both the increment operator and
constructor) to check if a newly constructed inner iterator is invalid
and, if so, to immediately increment until we get a valid one (or reach
the end).

This fixes the unit test failure.