Add function to read shared config
Add an overload of read_config_file to read one or more configuration
files (optionally merging contents) from standard shared data locations.
This can be used to read config files distributed with an application,
while providing a mechanism for the user to override settings in the
same.

For now, the caller is responsible for specifying the install prefix of
the executable using this function. This is because determining the same
is non-trivial, and because frankly the only implementation for doing so
that I fully trust is from Qt. (Fortunately, the immediate intended user
of this is the MAP-Tk GUI, which means we can use Qt for this.)