Add rgb_color comparison operators
Add operators == and != to rgb_color. These are helpful for trivially
determining if a color is equal to the default color, which can be used
to make a reasonable guess as to whether a PLY file provided colors or
not (e.g. if all colors are default, either it did not, or anyway the
colors are not meaningful).