Get color at image pixel
Add an additional accessor to vital::image to get the color at a
specified pixel as an rgb_color (which saves having to access each
channel value individually). See documentation for notes on assumptions
made w.r.t. the significance of image channels. (These assumptions are
expected to suffice for most cases.)