Support systems which do not have timer support.
The symbol VITAL_USE_CHRONO is set indicating if there is chrono/timer
support in the system. It is set based on the successful compilation
of test code but reset if we are building on VS12 (which seems to have
broken timer support)

Depending on the value of this symbol, the appropriate time include
file is configured to the build directory.