Improve enumerate_matrix tests
Add a test that iterators from the sparse matrix enumeration helper are
comparably distinct. This test is currently failing due to the
Eigen::SparseMatrix::InnerIterator lacking comparison operators(!), and
a naïve implementation of our own iterator's operator== assuming that
they exist. (The code compiles due to an implicit(!) conversion of the
InnerIterator to bool.) This can result in SEGV's when the helper is
used with the fallback implementation of VITAL_FOREACH, due to the
previous and incremented iterator erroneously comparing equal (thus
breaking the innermost loop of the implementation which relies on being
executed exactly once).

Also add a test against empty sparse matrices (which were initially
believed to be causing the problem).

See also http://eigen.tuxfamily.org/bz/show_bug.cgi?id=1192