Reverted change that tried to automatically reverse OpenCV channels
This hack was intented to fix issues with OpenCV defaulting to BGR color
space while everyone else expects RGB.  It worked, but didn't handle the
BGRA case and could be confusing for users not expecting channels to
automatically reverse.

Instead we should explicitly convert BGR to RGB where needed, such as in
the OpenCV image_io algorithm.