Use a faster method to get the Fully Qualified Domain Name
The previous approach was taking more than 30 seconds to get a FQDN on
some platforms.  The problem was that it iterated through each network
interface and did a domain lookup on each.  Some of these look-ups fail
but the timeout time is very long and non-configurable.

This alternate implementation is based on a suggestion on Stack Overflow:
http://stackoverflow.com/questions/504810/how-do-i-find-the-current-machines-full-hostname-in-c-hostname-and-domain-info

It uses getaddrinfo instead, which provides all the names upfront, so
don't need a look-up for each interface.