Remove check for impossible null reference.
Null references do not exist in C++ and we don't need to check for them.
This code was generating the following warning:

warning: reference cannot be bound to dereferenced null pointer in
well-defined C++ code; comparison may be assumed to always evaluate to
false

If there really is a case in exception handling where the reference is
null, it seems that should be handled elsewhere, e.g. before a NULL
pointer is converted to a reference.