Apply Necker reversal earlier in initialization
Waiting until the end to test for and correct Necker revsersal may
be too late to recover.  Reversal can't be detected if you try it too
soon.  This patch tries Necker reversal the first time the Necker
reversed initial RMSE is within the specified threshold of the
non-revsersed RMSE.  Whether we apply reversal or not, we only try it
once.