Enable thread safety in Cereal
Cereal is not thread safe by default, but provides a way to enable
thread safety at a small performance cost.  This commit enables thread
safety in Cereal by default.  Note that it is still not safe to access
one archive from multiple threads, but with this change it is safe to
access a different archive in each thread.