Add new geodesy types
Add "additional" geodesy types. These are intended to (eventually)
replace the existing lat/lon and UTM types, which must assume a CRS and
are less convenient to use. The new types are based on ViViA's geodesy
types, which define a geodetic location as a raw vector, and a CRS that
specifies how said vector relates to the Earth. This simplifies use by
having a single data type which does not impose a particular CRS or
style of representation and allows for more transparent on-demand
conversion between CRS's.