Added a string type and port trait in sprokit
Also reworded the description of the descriptor_set port trait since it
can and will be used for more than just feature-point descriptors. An
alternaltive is to name a separate type/port trait for other descriptor
types, but the existing one is general in name, so the description
should also be general.