"Implement" geo-conversion interface
Geo-conversion is tricky. Because we want caching to be integrated into
the geodetic types (to prevent wanton misuse), the conversion functions
must also be members of the types. However, the implementation of
conversion requires PROJ or similar, which is contrary to our goal of
the types having minimal to no dependencies.

To work around this, add a globally registered conversion functor. This
allows an implementation (e.g. an arrow) to register itself as the
implementation of conversion so that the types can call such an external
implementation. (If no conversion functor is registered, conversions
will just fail.)