Implement geodetic conversion in PROJ
Add an implementation of geo_conversion in the PROJ arrow, and have the
arrow set itself as the conversion functor on load. This makes geodetic
conversion possible as long as the PROJ arrow is available.

(We'll probably need to revisit this if/when there are multiple
implementations available, but this should be good for the time being.)