Start adding geo_polygon tests
Add geo_polygon tests, modeled after the geo_point tests. These are
incomplete, as the "api" test has not been updated from the point test
to work on polygons. However, the needed changes are making me
increasingly annoyed at the paucity of the home-grown test utilities,
and I am strongly leaning toward rewriting the test using Google Test.
This commit is a checkpoint to preserve the state where most of the test
works with the home-grown utilities, prior to wholesale conversion.