Add LL -> UTM/UPS helper
Add an additional helper function to the geodesy module to determine the
UTM/UPS zone of a latitude, longitude pair. This is needed to convert
coordinates to UTM, since it is impractical for the user to know the
appropriate UTM zone a priori.

Since UTM/UPS zones are based on a fairly simple system, it is plausible
to hard-code the implementation of this function, rather than needing to
rely on an external library. (This is similar to what is being done in
the old geo_map algorithm which this new functionality will eventually
replace.)