Make track constructors protected.
Track objects must always be held in a shared_ptr due to the
shared_from_this function.  Also we can't add or copy track states
in the constructor.  This change makes the constructor and copy
constructor protected and adds a public static factory function,
"make()", and a public clone function to work around these constraints.