Don't return rotation quaternion by reference
Change rotation::quaternion() to return by value, rather than returning
a reference to the internal data member. The latter is a subtle trap for
anyone calling the method on a temporary object; if the caller also
binds the return value to a reference, the resulting object will be
invalidated before it can be used. (In particular, this was causing many
of the VXL arrow tests to fail horribly because arrows/vxl/camera.cxx
was running afoul of this!)