Add helper for test temporary files
Add a test header providing a helper to obtain a name suitable for a
temporary file written by a test. This uses the known-"dangerous"
tempnam(), although it is less dangerous that simply hard-coding a file
name, which would not be unusual for a test to do.

Also update the config_block_io test to use this. Although that test was
previously using the "secure" mkstemp, it was using it in a way that
voided the security guarantees of mkstemp.

Note: The new helper is going to be used by another test. The motivation
for creating a helper was to improve consistency and avoid code
duplication.