Port last Ceres test to Google Test
Port the Ceres reprojection_error test to Google Test. This is the last
of the Ceres tests to be ported (and is, in fact, the last of the arrows
tests that are going to be ported at this time). This one was just
screaming to be parameterized, and has been, although this does result
in test names that are just a bit lengthy.