Assume std::chrono is available
Remove the configuration of the cpu_timer and wall_timer headers which
was present in case of older systems that lack std::chrono, since these
days we are assuming full availability of C++11. (This is also
consistent with the previous commit that turns lack of said features
into configure errors.) This will allow us to assume that these timers
are always available.

Also, fix a conditional check in wall_timer.h that provided a dummy
implementation unless a symbol was defined. This was redundant with the
configured-header approach, and worse, the symbol being checked was not
being defined anywhere, resulting in wall_timer never being available.