Minor improvements to Eigen test helpers
Include gtest.h via test_eigen.h. I was on the fence about doing this
since it isn't absolutely necessary, but the purpose of test_eigen.h is
to define additional Google Test assertions, so it makes sense to do so.
Also add and use static instances of the comparators, which produces
slightly better messages when an assertion fails.