Fix formatting of inline literals in documentation
Fix several reST documentation files that were using single-back-tick to
mark up literal text. This specifies "default interpreted text role",
which may happen to work in some contexts, but is not correct according
to plain reST, and results in the text being unstyled when viewed in
Github. Instead, use double-back-tick, which is the markup for inline
literals.