Export more complete pixel traits
Change the list of types for which we explicitly instantiate image pixel
traits to use the core type names, rather than cstdint type names. This
is necessary to ensure that both long and long long variants are
exported, which is important in case the cstdint types and VXL types are
not the same.

See also 3339294b3e82.