Implement alternate init string splitting
Implement an alternate mechanism to extract CRS properties from a PROJ.4
init string that does not rely on <regex>, as <regex> is broken before
GCC 4.9, and requiring that would be... awkward. (In particular, it is
the latest compiler available on CentOS 7, which we would really like to
support.)

Not squashing this in order to preserve the history of the <regex>
implementation.