Avoid loading plugins at build time
Tweak unit tests that need plugins to do so via a Google Test even
listener that fires at the start of test execution, rather than directly
in main(). This avoids attempting to load plugins when merely listing
tests, particularly when this happens at build time as part of test
discovery.