Fix config_block_set_value_cast<geo_polygon>
Fix algorithm to compute the correct number of decimal digits when
serializing a geo_polygon for output to a config_block. The previous
algorithm was failing to take the logarithm of the polygon magnitude,
such that we were setting the decimal digits precision to some screwy
value. For whatever reason, this apparently just happened to work on
Linux (possibly because it was producing a negative number that
libstdc++ ignored), but was causing Windows to output an empty string
instead of the coordinate value numbers.

Thanks go to the failing unit test for noticing that something was
wrong!