Remove use of VITAL_NOTHROW
Replace uses of VITAL_NOTHROW with noexcept. This is a no-op change,
since VITAL_NOTHROW is defined to noexcept anyway, but it is part of an
effort to modernize the code by by removing use of C++11 feature shims
which existed for compatibility with older compilers (which we no longer
support).