Remove use of VITAL_DEFAULT_[CD]TOR
Replace uses of VITAL_DEFAULT_CTOR and VITAL_DEFAULT_DTOR with
`= default`. This is a no-op change, since those macros expand to the
same, but it is part of an effort to modernize the code by by removing
use of C++11 feature shims which existed for compatibility with older
compilers (which we no longer support).