Refactored process and scheduler factories.
Created an abstract base class and derived tow concrete factory classes.
One for cpp {processes,schedulers} and one for python. The details
of converting python object into a process shared_ptr are delegated
to the derived classes.

This approach solves the problem in the plugin_explorer where there
was an additional factory type for the pybind11::object that was not
handled.