Add set of persistent object properties to the process class.
Added process properties that are bound to the process
object. Process properties were created based on the process class
hierarchy, where the derived process would create the set of
properties. There are cases where the derived process does not know
all the properties, such as when process instrumentation is
enabled. The solution is to add a set of properties that is part of
the process object that can accumulate these condition based
properties.