Fix file name collisions in OpenCV test
Change OpenCV feature_descriptor_io test to use temp_file_name to
generate unique names for the files the various tests write, rather than
using hard-coded names (which are the same across different tests!).
This avoids test failures when the tests are run in parallel due to the
various tests all trying to write to the same file.