Fix logger "sameness" assertion
Relax assertion that requesting a logger by name multiple times gives
back the same pointer, as this is apparently not true for all
implementations. Instead, restore the pre-#311 assertion that the second
reference has the same log level, and extend this to verifying that the
second reference's log level changes when the log level is changed via
the first reference.