Fix preprocessor UB in OpenCV test
Refactor how algorithm type-paramaterized tests are set up in order to
avoid use of preprocessor directives inside of a macro's argument list.
Although GCC and clang do the "right" (or at least, the expected and
desired) thing here, MSVC basically croaks on the old code. According to
the standard¹, [cpp.replace]¶11, we are in Undefined Behavior land.

¹ Section 16.3 as of N3337
  (http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2012/n3337.pdf)