Add timeout to Google Test discovery
Tweak Google Test discovery to abort if the test executable fails to
complete within 30 seconds. This is a work-around for the thread_pool
test which is currently hanging on our Windows dashboard build for
reasons not-yet-known. (Curiously, it only hangs when given
--gtest_list_tests; when the tests actually run, they pass, and the
executable exits normally.)

This is not currently upstream, but I intend to submit a version of this
for inclusion in CMake 3.10.1.