Add fields_at_rows() to return the active fields for a set of rows.
The fields_at_row() method returns the active fields for a single
row. This commit adds a fields_at_rows() method to scan a set
of rows more efficiently than one-at-a-time.

(It was originally implemented to speed up track_reader_example
for KPF, but then I found a wholly different KPF bug that was making
things slower. I had already written this, though.)