Make thread_pool::available_backends static
Change thread_pool::available_backends to be a static method. There
really is no reason that this needs an instance of the thread_pool, and
requiring one can get... awkward for the unit test, which needs the list
of available backends in order to set up the test cases.

In particular, this works around, but does not really solve, #407, and
should be sufficient to get the Windows builds working (at least with
respect to this particular issue).