KPF YAML polygon I/O hack.
Back before YAML, the text-based KPF output wrote out polygons like this:

  poly0: 3 100 100 200 200 300 300

...where the '3' was the number of (x,y) points. However, the naive yaml
output:

   poly0: 3 [[ 10, 10 ],[ 20, 20 ],[ 10, 30 ],],

...is ill-formed. We just need to not emit the number of points:

   poly0: [[ 10, 10 ],[ 20, 20 ],[ 10, 30 ],],

...however, the upstream token parser REQUIRES the number of points.

The correct fix is to rework the upstream token parser to have more structure.

The quick hack fix is to (a) detect when we're parsing a poly and (b) insert
the number of tokens into the stream.