Fix per-configuration built-in plugin paths
Modify how we set up built-in plugin paths to perform substitution of
the `$<CONFIGURATION>` CMake generator expression in the C++ source
code. This is necessary because the configuration is not the final path
component (so we can't just tack on the configuration name to handle
per-configuration paths at the C++ level), and, since we can't pass the
list of paths directly as a defined symbol due to argument quoting
issues, because we are obtaining the paths through a configured file,
which cannot easily be made to contain per-configuration information.

This (along with the previous commit) should fix plugin loading on
multi-configuration generators (particularly Visual Studio) when relying
on the built-in paths, which in particular should fix the several dozen
tests that are currently failing because their needed plugins are
missing.