Expanded video_input API to support frame seeking.
Expanded the the video_input API to support frame seeking. Added
two methods, seekable() which just returns true if the video input
supports seeking and seek_frame() which seeks to the requested frame.
Stubbed out the methods for all the implementations so the code still
compiles and tests still pass.