Modify embedded pipeline to allow config update before running.
In order to support inspecting, verifying and modifying the pipeline
config after the config has been ingested but before the pipeline has
been built, a method is added to allow derived classes to access the
config at the correct point in the execution.

This is useful in cases where the pipeline description must be read from
a file, but it must also meet specific requirements imposed by the
application. A derived class can implement the update_config() method
and verify that the requirements have been met, or add required entries
to the config before building the pipeline.