Added frame_skip option to ffmpeg implementation.
Added frame_skip option for ffmpeg implementation of video_input.
When next frame is called it will skip the number of frames that
are set in frame_skip. num_frames return value only returns the
number of frames that wouldn't be skipped. The metadata_map only
includes entries for frames that won't be skipped. seek_frame
returns false if the requested would be normally be skipped.