Catch metadata exceptions, log a warning, and continue.
It is usually better to catch exceptions due to bad metadata,
discard the bad metadata, and move on to the next metadata rather
let the program crash or catch the exception at a higher level.

This change prevents tools like dump_klv from crashing when encountering
a bad metadata packet.