Add method to get current video input time stamp
Add new method frame_timestamp to video_input interface to retrieve the
current time stamp (i.e. the one that was produced by the most recent
call to next_frame). This is useful for consumers that need to reference
the time stamp more than once, especially if storing it themselves is
inconvenient.