Ensure that export variables are configured as proper lists
Since the file `kwiver-config.cmake` is generated at build time by
`kwiver_configure_file` macro, the list variables were not being parsed
correctly. The lists were being truncated to the first value before
being replaced into the config file.

This change fixes this. The basic idea is handing the list as a string
for `configure_file` for @ONLY replacement. The list is then rebuilt in
the config file.