Introduce range helpers
"Ranges" is a feature space of upcoming C++ (i.e. C++17 and later) that
simplifies common operations when working with sequences. Since it will
be some time before KWIVER can rely on native support, this starts
implementing some of the more useful bits ourselves.

As of this commit, integer sequences (which are useful for certain
classes of template functions) and a sliding window view are
implemented. The former is needed by the latter. The latter is wanted by
a unit test.

Upcoming commits will add additional features, including filtering and
transforming.