Make template adapters work on C-style arrays
...sort of. Template adapters, e.g. sliding, don't work on C-style
arrays prior to this commit, since it would require overloading
operator| for two pointer-type operands, which is not permitted (see
https://stackoverflow.com/questions/49638281). Unfortunately, right now,
I don't know of any way to make this work.

As a work-around, tweak the code so that it works if the adapter
function is invoked (i.e. `sliding<N>()` instead of `sliding<N>`). This
is a little aesthetically-challenged, but at least allows adapters to be
used as the first operation on a C-style array.