Allow default-construction of range iterators
Modify the range iterators to allow default construction and assignment.
The latter requires taking copies of the relevant functions for filter
and transform, and the former naturally produces a broken iterator
(though this is generally always the case for any iterator type), but
these changes are necessary to allow some combinations of chaining.