Generalize transform adapter function type
Modify transform_view to take the function type as a template parameter
rather than coercing the function into a std::function. This should
improve performance somewhat for certain functors, especially simple
function pointers. (This is more important since we now need to copy the
functor due to the previous commit.)