Force copies of range adapters
Create a reference-holder class for range adapters to use to hold the
"reference" to the input range. (This also allows some simplification of
how we obtain iterators.) Force this to take a copy if the input range
is itself a range adapter.

Copies of range adapters should be reasonably inexpensive, since they
are just holding iterators, functions, and the like. Taking a copy is
necessary to prevent intermediate adapters from being destroyed while
building an adapter chain. (Maybe one day C++ will solve lifetime
extension, and this will stop being a problem, but for now it's a
necessary evil.)