Refactor function-type range adapter macros
Since 8e98df2689e2, there is no longer a difference in the utility
classes to construct a transform_view or a filter_view, and this will
probably hold true for other function-type views as well. Thus, refactor
the utility classes to make their definitions part of the implementation
helper macro, rather than requiring them to be defined per adapter.