Adjust the exception hierarchy to easily catch any exception.
The name of the base vital exception is shortened to be memorable and
easy to use for catching any kwiver or vital exception. There has not
been any standardization on how to catch exceptions (or which one to
catch).

Change config block exceptions to be derived from the same base class
as the others. This change caused the exceptions to be grouped in a
separate library to remove a circular dependency vital <-> vital_config.