Simplify range iterator type helper
Refactor the helper to obtain range iterator types to use a detail
namespace, into which all of std is imported. This is a little uglier in
some senses than doing the import into a function, but avoids needing to
obtain the return type of a lambda, which apparently doesn't work with
Visual Studio. (The resulting code is also markedly less esoteric, if
less elegant.)