Use vital::uid for track_oracle track_uuid
Change data type of track_oracle's track_uuid to kwiver::vital::uid,
rather than std::string. This was originally a boost::uuid_t and was
changed to std::string in order to remove the boost dependency at a time
before vital::uid existed. However, since the intent is to hold a UUID,
and vital::uid is intended to be used for this purpose, it is a more
appropriate type for the field.