Introduce validity range adapter
Create a range adapter to filter a range for valid items. This is
effectively the same as the filter adapter, except the filter function
is fixed as applying double-negation to the item. This is especially
convenient for iterating over a collection of pointers, in order to
visit only non-null pointers.