Relevancy/confidence in 19th column should apply to frames, not tracks.
As written, when a 19th column is present in kw18, that value is
only applied to the track, not to the individual frames. So if you
have a track with three frames:

 track  frame   .... column-19

  1      10     ....  0.5
  1      11     ....  0.6
  1      12     ....  0.7

...then the relevancy will be assigned to track 1, and the value
will be the last one read (here, 0.7); the values of 0.5 and 0.6
are lost.

This fix applies the relevancy to the frames, not the track.

Since essentially the only use of having a 19th column in a kw18 is
to assign frame-level confidence values, and this is what the scoring
code expects, this will fix detection-mode scoring and shouldn't break
anything.