Force masks to binary with a threshold
Assuming mask image is intended to have values of only 0 and 255,
we can force masks into these values with a threshold at 128.  This is
important for compressed image sources where the values may not have
exactly the expected values.  The OpenCV code considers only 0 values
to be masked.  If the low values do not actually hit 0 then the mask
will not apply correctly.