Changed return type of static protobuf converters to void.
The return code was removed because it was never set to anything but
the default and not checked by caller. If there is an error in these
low level converters, it should be thrown as an exception and caught
at the process level. There is no way to recover from a conversion
failure at these low levels.