Moved convert protobuf() functions to a separate file.
Locating the convert_protobuf() functions with the serializer
implementation is not scaling well. There are cases where it is
convenient to have protobuf's that are not first class pipeline
data types. Putting those in the converter class as static methods
is causing problems finding which class has the specific converter
you are looking for. Moving them all into one name space simplifies
the code.