Fix usage of matplotlib in aggregate_plots
The aggregate graph wasn't correctly setting the upper bound of the
y-axis to be the maximum of all the fish data in the graph. It was
only setting it for the first species listed because of when we
were calling ax.set_ylim(). This change moves it to the very end so
that all of the data is taken into account when it's called.