Add process method to handle undefined ports.
This method supports dynamically creating ports as they are
specified making connections. The previous approach to dynamically
creating ports was to override the _*_port_info() method. While
this approach works, the _*_port_info() call is also called when moving
data over an edge. The derived implementation of this method can
add considerable overhead if it is not carefully programmed. In
addition, it is not obvious that the _*_port_info() calls are part of
the data transfer operations, so efficiency may not be part of the
design.

These new methods are only called when a port is being requested of a
process and it is not defined. They are never called again after the
port has been defined, eliminating the need to highly optimize these
methods.