Add YOLOv3 seed weights and model
The weights are simply those available from the official
site (https://pjreddie.com/media/files/darknet53.conv.74), and the
configuration file is lightly adapted from the version in the Darknet
repository as of now, December 2018 (commit 61c9d02).

Most notable is the shorter training to match a similar reduction in
Viame's included YOLOv2 model, as well as the disabling of random
network resizing, which effectively disregards the set width and
height.