Create EmbeddedPipelineWorker
Create a Qt utility class to simplify embedding a KWIVER pipeline inside
of a Qt application. (Note that, as a Qt-based class, this follows Qt
naming conventions, which differ from the conventions used elsewhere in
KWIVER.)

This code is largely lifted from ViViA, with the objective that ViViA's
version can eventually become a subclass of this class. It will also be
used for an "example" pipeline viewer which will live in KWIVER proper.