Improve style in arrows/qt
Apply several style changes found with uncrustify. This is not the exact
output of uncrustify, as it has fits with the copyright headers (which
are over line length and have other issues that ought to be addressed at
some point, but not now) and a few bugs where it is trying to apply
incorrect formatting.