Also show text on pipeline viewer icons
Tweak pipeline viewer to always show text next to the toolbar icons.
This is helpful because, rather than fuss with supplying our own icons,
we took the easy route of relying on the system icon theme... which
works great (and is TRTTD) on XDG-compliant platforms, and not so well
on others (e.g. Windows, macOS). Always including the text may not be as
pretty, but at least won't have users scratching their heads wondering
about the blank buttons.