more ranges (#721)
* Fix wrong-order test assertions

Fix order of arguments in the filter range adapter test. (These were in
the "natural" order actual, expected, but the macro wants expected,
actual... As this feels less natural, it is not too surprising that
these were backwards.)

* Introduce indirect range adapter

Create a range adapter to apply a level of indirection to a range during
iteration. This effectively suppresses dereferencing of the iterator, so
that the user sees the iterators rather than the values, while still
being able to use the more terse and convenient syntax of a range-based
for. Also, fix some minor lingering issues with creating a mutating
range adapter.

* Introduce validity range adapter

Create a range adapter to filter a range for valid items. This is
effectively the same as the filter adapter, except the filter function
is fixed as applying double-negation to the item. This is especially
convenient for iterating over a collection of pointers, in order to
visit only non-null pointers.

* Update release notes to mention new range helpers

Mention new range helpers added in the previous commits in the release
notes. Also, get rid of all the trailing whitespace in the release
notes, and fix an over-length line.

* Add additional range tests

Add a test that range::filter and range::valid work on empty input
ranges. Add a test that range::valid works on an input range with no
valid items.

* Don't crash on empty ranges

Fix issue with range::filter and range::valid that could cause a crash
if the input range is empty. (It seems that the compiler optimizations
are aggressive enough that the problematic code gets squelched in
release builds, so the problem only shows up in a debug build, and even
then, only when running the tests under valgrind.)

* Work around buggy ADL in VS

Move the range operators to the global namespace. This works around
buggy ADL in visual studio that apparently does not consider the
namespace of the RHS if the RHS is a function pointer. (GCC and Clang
don't have this bug.)

* misc range fixes